server/requirements.txt
anthropic
flask
flask-cors
beautifulsoup4
chromadb
sentence-transformers
python-dotenv